+++ src/channels/index.ts
@@ -3,6 +3,7 @@
  */
 
 export * from './types.js';
+export * from './setup.js';
 export * from './telegram.js';
 export * from './slack.js';
 export * from './whatsapp/index.js';
